pyagent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyagent"
version = "0.1.0"
description = "A lightweight AI Agent framework with MCP and Skills support"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PyAgent Team" }
]
keywords = ["ai", "agent", "mcp", "llm", "skills"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "anyio>=4.0.0",
    "typing-extensions>=4.0.0",
    "pyyaml>=6.0",
    "watchfiles>=0.20.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.25.0"]
zhipu = ["zhipuai>=2.0.0"]
ui = ["gradio>=4.0.0"]
vector = ["chromadb>=0.4.0", "sentence-transformers>=2.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "pyagent[openai,anthropic,zhipu,ui,vector,dev]",
]

[project.scripts]
pyagent = "pyagent.cli:main"

[project.urls]
Homepage = "https://github.com/example/pyagent"
Documentation = "https://github.com/example/pyagent#readme"
Repository = "https://github.com/example/pyagent"

# Hatch build configuration
[tool.hatch.build.targets.wheel]
packages = ["src/pyagent"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
